+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) for the `dbcrossbar` CLI tool. (The `dbcrossbarlib` crate is an internal-only dependency with no versioning policy at this time.)
 
+## [Unreleased]
+
+### Fixed
+
+- gs: Retry `delete` operations after 403 Forbidden errors. This appears to be part of the same hypothesized permissions race that affected `extract` in v0.5.0-beta.3.
+
 ## [0.5.0-beta.3] - 2021-12-31
 
 ### Fixed